If 2 fix

+++ 5-if2/shoot-8.cpp
@@ -2,7 +2,7 @@
 using namespace std;
 
 int main() {
-  int x, y;
+  long long int x, y;
   cin >> x >> y;
   unsigned int circle = x*x + y*y;
   cout << ((y > x*x - 6 && circle < 16) ||
